Move the Dutch holidays in a dedicated file
